[EASI-3788] IT Gov - move v1 components to v2 folder (#2345)
* Move LcidInfo to v2 folder

* Move RequestDecision folder
